backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
python-dotenv>=1.0.0
requests>=2.31.0
PyJWT>=2.8.0
cryptography>=41.0.0
langchain>=0.1.0
openai>=1.0.0
playwright>=1.40.0
pytest>=7.4.0
pytest-django>=4.7.0
black>=23.12.0
ruff>=0.1.9
mypy>=1.7.0
python-docx>=1.1.0
Pillow>=10.1.0

